feat: organize code into internal & external
