+++ internal/external/aws.go
@@ -1,4 +1,4 @@
-package internal
+package external
 
 import (
 	"io"
