+++ internal/external/aws_test.go
@@ -1,4 +1,4 @@
-package internal
+package external
 
 import (
 	"os"
